SW-4871 Add support for draft sites in BoundariesAndZones component
- show zones/subzones in map
- show zones/subzones in list views
- extracted out GenericZoneView from PlantingSiteZoneView for reuse across draft sites and normal planting sites
- cleaned up some selector types
- enabled search in the zone/subzone views (for some reason this was using a hardcoded '' as search query)
