+++ src/scenes/PlantingSitesRouter/view/PlantingSiteDetailsTable.tsx
@@ -97,7 +97,7 @@ const DetailsRenderer =
       const url = zoneViewUrl
         .replace(':plantingSiteId', plantingSiteId.toString())
         .replace(':zoneId', row.id.toString());
-      return <Link to={url}>{row.name as React.ReactNode}</Link>;
+      return <Link to={url}>{(row.name || '--') as React.ReactNode}</Link>;
     };
 
     if (column.key === 'name') {
